requirements.txt
feedparser>=6.0.10
beautifulsoup4>=4.12.2
requests>=2.31.0
PyYAML>=6.0.1
Jinja2>=3.1.2
schedule>=1.2.1
lxml>=4.9.3
python-dateutil>=2.8.2
